pyproject.toml
[project]
name = "photoutils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Tom Shafer", email = "tom@tshafer.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pyexiftool>=0.5.6",
    "typer>=0.15.4",
    "watchdog>=6.0.0",
]

[dependency-groups]
interactive = [
    "jupyter>=1.1.1",
]

[project.scripts]
photoutils = "photoutils.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
ignore_missing_imports = true
